First version of remote execution framework
This framework uses a simple technique to clone itself
over an SSH connection and establish a SSH tunnel to itself
with the explicit goals that rpc can be excuted between the client
and the remote server. This makes it easier for us to write the
code in pure golang and execute the code remotely and not using
shell command to execute remotely.

Please give it a try the rpclcient and common files in the cmd module
contains a simple example RPC client server code which can be replaced
with something more interetesting and real for your project.